requirements.txt
dash
dash-ag-grid
pandas
pyodbc
python-dotenv
tabulate